[BackEnd] Refactored
