+++ UserProfileService/application/service/user-service.js
@@ -2,9 +2,8 @@ require('dotenv').config()
 const bcrypt = require('bcrypt')
 
 const ormUser = require('../orm/user-orm')
-const { Response } = require('../../util/response')
-const { STATUS_SUCCESS, STATUS_FAIL } = require('../../util/enums')
-const { InternalServerError, MissingArgsError } = require('./common')
+const { STATUS_SUCCESS, STATUS_FAIL } = require('../util/enums')
+const { Response, InternalServerError, MissingArgsError } = require('./common')
 
 exports.CreateUser = async (req, res) => {
   try {
