+++ UserProfileService/application/util/enums.js
