+++ UserProfileService/controller/routeController.js
@@ -2,7 +2,7 @@ const { Router } = require('express')
 
 const UserService = require('../application/service/user-service')
 const AuthService = require('../application/service/auth-service')
-const MatchService = require('../application/service/match-service')
+// const MatchService = require('../application/service/match-service')
 
 const routes = Router()
 
@@ -14,11 +14,11 @@ routes.get('/user/auth', AuthService.AuthenticateToken, AuthService.AuthRoute)
 routes.post('/user/token', AuthService.RefreshToken)
 
 // TEMP: For Testing Purposes only
-routes.get('/match/find/:username', MatchService.FindUserMatched)
-routes.post('/match/create', MatchService.CreateMatch)
-routes.delete('/match/delete', MatchService.DeleteMatch)
-routes.post('/match/find', MatchService.FindMatch)
-routes.delete('/match/delete-expired', MatchService.DeleteExpiredMatches)
+// routes.get('/match/find/:username', MatchService.FindUserMatched)
+// routes.post('/match/create', MatchService.CreateMatch)
+// routes.delete('/match/delete', MatchService.DeleteMatch)
+// routes.post('/match/find', MatchService.FindMatch)
+// routes.delete('/match/delete-expired', MatchService.DeleteExpiredMatches)
 // ---
 
 /**
