+++ UserProfileService/test/stubs.test.js
