--- UserProfileService/util/enums.js
@@ -1,2 +1,0 @@
-exports.STATUS_SUCCESS = 'success'
-exports.STATUS_FAIL = 'fail'
